+++ src/value.c
@@ -11,7 +11,7 @@ void init_value_array(ValueArray* array) {
 
 void write_value_array(ValueArray* array, Value value) {
     if(array->capacity < array->count + 1) {
-        int old_capacity = array->capacity;
+        size_t old_capacity = array->capacity;
         array->capacity = GROW_CAPACITY(old_capacity);
         array->values = GROW_ARRAY(
             Value,
